+++ code/sparkngin/build.gradle
@@ -8,6 +8,8 @@ dependencies {
   compile group: 'io.vertx', name: 'vertx-platform', version: '2.1RC2'
   compile group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.3'
 
+  compile group: 'com.jcraft', name: 'jzlib', version: '1.1.3'
+
   testCompile group: 'io.vertx', name: 'testtools', version: '2.0.3-final'
 }
