+++ code/sparkngin/src/test/java/com/neverwinterdp/sparkngin/jetty/impl/MessageServlet.java
@@ -1,4 +1,4 @@
-package com.neverwinterdp.sparkngin.jetty;
+package com.neverwinterdp.sparkngin.jetty.impl;
 
 import java.io.IOException;
 import java.io.InputStream;
