+++ code/sparkngin/src/test/java/com/neverwinterdp/sparkngin/vertx/EventBusPingUnitTest.java
@@ -7,6 +7,8 @@ import org.vertx.java.core.Vertx;
 import org.vertx.java.core.eventbus.Message;
 import org.vertx.java.platform.Verticle;
 
+import com.neverwinterdp.sparkngin.vertx.impl.EmbbededVertxServer;
+
 public class EventBusPingUnitTest {
   @Test
   public void testEventBusPingVerticle() throws Exception {
